requirements.txt
meross_iot==0.4.10.4
pyyaml>=6.0
httpx>=0.27
